# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aithos Agents API",
    "version": "0.1.0",
    "description": "Phase 1: public agent creation, one owner key, Markdown skills, remote MCP connectors and asynchronous tasks. One global model budget; no A2A invocation transport yet."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3188",
      "description": "Ephemeral local mode; metadata only unless Bedrock explicitly enabled."
    },
    {
      "url": "https://{apiId}.execute-api.{region}.amazonaws.com",
      "description": "Deployed API Gateway HTTPS endpoint",
      "variables": {
        "apiId": {
          "default": "YOUR_API_ID"
        },
        "region": {
          "default": "eu-west-3"
        }
      }
    }
  ],
  "security": [
    {
      "OwnerKey": []
    }
  ],
  "paths": {
...
    "/v1/agents": {
      "post": {
        "operationId": "createAgent",
        "summary": "Create an agent and return its one-time owner key",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCreated"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentInput"
              }
            }
          }
        }
      }
    },
    "/v1/agents/{agent}": {
      "get": {
        "operationId": "getAgent",
        "summary": "Read agent configuration",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/agent"
          }
        ]
      },
      "patch": {
        "operationId": "updateAgent",
        "summary": "Replace agent name and description",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "description": "Name is required. Omitted description resets it to an empty string.",
        "parameters": [
          {
            "$ref": "#/components/parameters/agent"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentInput"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteAgent",
        "summary": "Disable agent and clear private records",
        "responses": {
          "204": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/agent"
          }
        ]
      }
    },
    "/v1/agents/{agent}/key": {
      "post": {
        "operationId": "rotateOwnerKey",
        "summary": "Rotate the owner key and revoke the old key",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerKey"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/agent"
          }
        ]
      }
    },
    "/agents/{agent}/agent-card.json": {
      "get": {
        "operationId": "agentCard",
        "summary": "Public card with custom REST binding",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCard"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "description": "A2A Agent Card structure. Phase 1 exposes a custom REST interface, not an A2A JSON-RPC, gRPC or streaming endpoint.",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/agent"
          }
        ]
      }
    },
    "/v1/agents/{agent}/skills": {
      "get": {
        "operationId": "listSkills",
        "summary": "List private skill configuration",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Skills"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/agent"
          }
        ]
      }
    },
    "/v1/agents/{agent}/skills/{skill}": {
      "put": {
        "operationId": "saveSkill",
        "summary": "Create or replace a Markdown skill",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Skill"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/agent"
          },
          {
            "$ref": "#/components/parameters/skill"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillInput"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteSkill",
        "summary": "Remove a skill",
        "responses": {
          "204": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/agent"
          },
          {
            "$ref": "#/components/parameters/skill"
          }
        ]
      }
    },
    "/v1/agents/{agent}/connectors": {
      "get": {
        "operationId": "listConnectors",
        "summary": "List connectors without secrets",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connections"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/agent"
          }
        ]
      },
      "post": {
        "operationId": "createConnector",
        "summary": "Add a remote MCP connector",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connection"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "description": "Stores configuration. This does not verify the remote endpoint or complete OAuth consent.",
        "parameters": [
          {
            "$ref": "#/components/parameters/agent"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionInput"
              }
            }
          }
        }
      }
    },
    "/v1/agents/{agent}/connectors/{connection}": {
      "delete": {
        "operationId": "disconnectConnector",
        "summary": "Disable connector and remove local credentials",
        "responses": {
          "204": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "description": "Does not call a provider revocation endpoint or undo previously dispatched actions.",
        "parameters": [
          {
            "$ref": "#/components/parameters/agent"
          },
          {
            "$ref": "#/components/parameters/connection"
          }
        ]
      }
    },
    "/v1/agents/{agent}/connectors/{connection}/authorize": {
      "post": {
        "operationId": "authorizeConnector",
        "summary": "Start or restart OAuth consent",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthStart"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/agent"
          },
          {
            "$ref": "#/components/parameters/connection"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthInput"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "OwnerKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "agent owner secret",
        "description": "Single key for agent management and invocation."
      }
    },
    "schemas": {
...
      "AgentCreated": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "card_url": {
            "type": "string",
            "format": "uri"
          },
          "owner_key": {
            "type": "string",
            "writeOnly": false,
            "description": "Returned once; save securely. Also permits invocation."
          }
        },
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "description",
          "created_at",
          "card_url",
          "owner_key"
        ]
      },
...
    },
...
    "responses": {
      "Error": {
        "description": "Application error or request extraction error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "text/plain": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}